- off_t in freebsd is 64 bit, so a temporary fix is to cast as long   int. This means dialogs with size information will be incorrect   in freeBSD for files greater than 4GB.
(Old svn revision: 542)

+++ xftree/entry.h
@@ -24,7 +24,7 @@
 
 #ifndef __ENTRY_H__
 #define __ENTRY_H__
-
+#include <sys/types.h>
 typedef struct
 {
   int year;
@@ -39,7 +39,7 @@ typedef struct
 {
   gchar *path;
   gchar *label;
-  int size;
+  off_t size;
   int type;
   int flags;
   int mtime;
